Refactor to remove use of StatefulSetLists
A Logstash instance will only have one StatefulSet, so there is no
need to use StatefulSetLists
